requirements.txt
torch
torchvision
torchaudio
ultralytics
opencv-python
matplotlib
pycocotools
tqdm
typer
rich
pyyaml
notebook
jupyter
ipykernel
nbformat
nbconvert
